feat(ffi): winscard: implement `cancel` and `list_reader_groups` function for system scard context;
Improved multistring parsing.
