+++ ffi/src/winscard/system_scard/card.rs
@@ -10,6 +10,7 @@ use winscard::winscard::{
 };
 use winscard::{Error, ErrorKind, WinScardResult, CHUNK_SIZE};
 
+use super::parse_multi_string_owned;
 use crate::winscard::buf_alloc::SCARD_AUTOALLOCATE;
 
 pub struct SystemScard {
@@ -55,9 +56,9 @@ impl WinScard for SystemScard {
             })?;
 
             let readers = if let Ok(readers) =
-                parse_multi_string(unsafe { from_raw_parts(reader_name, reader_name_len.try_into()?) })
+                parse_multi_string_owned(unsafe { from_raw_parts(reader_name, reader_name_len.try_into()?) })
             {
-                readers.iter().map(|&r| Cow::Owned(r.to_owned())).collect()
+                readers
             } else {
                 try_execute!(unsafe { pcsc_lite_rs::SCardFreeMemory(self.h_card_context, reader_name as *const _) })?;
 
@@ -264,14 +265,4 @@ impl WinScard for SystemScard {
         }
         Ok(())
     }
-}
-
-fn parse_multi_string(buf: &[u8]) -> WinScardResult<Vec<&str>> {
-    let res: Result<Vec<&str>, _> = buf
-        .split(|&c| c == 0)
-        .filter(|v| v.is_empty())
-        .map(|v| std::str::from_utf8(v))
-        .collect();
-
-    Ok(res?)
 }
